requirements.txt
beautifulsoup4==4.12.2
boto3==1.34.0
matplotlib==3.10.1
pandas==2.2.3
pytest==8.4.2
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.44.1
